Move express-db functions to a db api file
