+++ urls.js
@@ -9,6 +9,7 @@ const request = require('request');
 const i18n_utils = require('./i18n/utils');
 const {migrate, connection, Map, History, Category} = require("./db/modelsB.js");
 const query = require("./db/query.js");
+const dbapi = require("./db/api.js");
 const sharp = require('sharp');
 // Local units
 const wizard = require('./units/wizard');
@@ -731,83 +732,11 @@ module.exports = function(app, apicache) {
         return records.filter(record => record.model === modelName)
     }
 
-    /**
-     * 
-     * @param {*} req 
-     * @param {*} res 
-     * @returns {Object} with exists flag, or a message 
-     */
-    async function admin_api_get_map (req, res) {
-        if (req.query.path) {
-            const map = await query.mapByPath(req.query.path, req.query.id);
-            res.send({
-                "exists": map ? true : false
-            });
-        }
-        else {
-            res.status(400).send({"message": "Invalid request"})
-        }
-    }
-
-    async function admin_api_get_categories (req, res) {
-        /** Get all Category records **/
-        const categories = await query.getAllCategories();
-        /** Remap to consume on wizard category search */
-        res.send(categories.map(category => new Object({"title": category.name, "id": category.id})));
-    }
-
-    /**
-     * Update multiple records.
-     * 
-     * @param {*} req 
-     * @param {*} res 
-     */
-    async function admin_api_action_update (req, res) {
-        /** Save all Map records **/
-        const mapRecords = getRecordsForModel(req.body.records, 'map');
-        // console.log(mapRecords);  // DEBUG
-        const updatedRecordsMessage = await updateRecordList(Map, mapRecords);
-        /** Save all Category records **/
-        const categoryRecords = getRecordsForModel(req.body.records, 'category');
-        // console.log(categoryRecords);  // DEBUG
-        const updatedCategoriesMessage = await updateOrAddCategories(categoryRecords);
-        /** merge errors on response **/
-        res.send({
-            error: updatedRecordsMessage.error || updatedCategoriesMessage.error, 
-            updateNumber: updatedRecordsMessage.count + updatedCategoriesMessage.count
-        });
-    }
-
-    function admin_api_action_undelete (req, res, published=true) {
-        admin_api_action_delete(req, res, published);
-    }
-
-    function admin_api_action_delete (req, res, published=false) {
-        // move to Draft
-        console.log(req.body);
-        Map.update(
-            { published: published }, /* set attributes' value */
-            { where: { id: req.body.id }} /* where criteria */
-          ).then(([affectedCount, affectedRows]) => {
-            // Notice that affectedRows will only be defined in dialects which support returning: true
-            // affectedCount will be n
-            res.send({error: false, updateNumber: affectedCount});
-          });
-    }
     // Enable json for express (to get req.body to work)
     app.use(express.json());
 
 
-    app.get('/admin/api/get/:name', function (req, res) {
-        let fun = eval('admin_api_get_' + req.params.name);
-        try {
-            fun(req, res);
-        }
-        catch (e) {
-            // Function not found, pass
-            res.send("Error")
-        }
-    });
+    app.get('/admin/api/get/:name', dbapi.adminApiGetName);
 
     /**
      * Internal admin API to C-U- (not Read, not real Delete) maps.
@@ -815,16 +744,7 @@ module.exports = function(app, apicache) {
      * @param  {Express response} res
      * @return {[type]}     [description]
      */
-     app.put('/admin/api/:action', function (req, res) {
-        let fun = eval('admin_api_action_' + req.params.action);
-        try {
-            fun(req, res);
-        }
-        catch (e) {
-            // Function not found, pass
-            res.send("Error");
-        }
-    });
+     app.put('/admin/api/:action', dbapi.adminApiAction);
 
     /**
      * Display Landing page
